Rename routes to api.
