+++ src/lib.rs
@@ -3,5 +3,5 @@ pub mod configuration;
 // pub mod models;
 pub mod error_handling;
 pub mod persistence;
-pub mod routes;
+pub mod api;
 pub mod telemetry;
